Separate `ctlpaneltype` into own module
